pyproject.toml
[project]
name = "pythonproject5"
version = "0.1.0"
description = ""
authors = [
    {name = "herman_shyrkavets",email = "gsiruovec@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "dependency-injector (>=4.48.3,<5.0.0)"
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "aioresponses>=0.7.6",
    "httpx>=0.27.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
